Add files via upload
The dataset contains information about daily trading activities for this company, including stock prices, trading volume, turnover, and other relevant metrics. Here's a breakdown of the columns in the dataset:

Date: The date of the trading activity, which represents the specific day when the data was recorded.

Symbol: The stock symbol or ticker symbol that uniquely identifies the company in the stock market. In this case, it is "MUNDRAPORT."

Series: The trading series, which typically indicates the type of security being traded. In this dataset, it is "EQ," which stands for equity shares.

Prev Close: The closing price of the stock on the previous trading day.

Open: The opening price of the stock on the current trading day.

High: The highest price at which the stock traded during the trading day.

Low: The lowest price at which the stock traded during the trading day.

Last: The last traded price of the stock during the trading day.

Close: The closing price of the stock on the current trading day.

VWAP (Volume Weighted Average Price): The average price at which the stock was traded during the trading day, weighted by trading volume.

Volume: The total number of shares of the stock traded on the current trading day.

Turnover: The total value of the stock traded on the current trading day, often measured in monetary units (e.g., Indian Rupees).

Trades: The total number of trades or transactions that occurred for the stock on the current trading day.

Deliverable Volume: The volume of shares that were physically delivered to the buyers on the current trading day.

%Deliverble: The percentage of deliverable volume relative to the total volume traded.

This dataset is valuable for analyzing the historical performance and trading patterns of the company's equity shares. It can be used for various financial analyses, including technical analysis, trend analysis, and the development of trading strategies. Additionally, it provides insights into trading volume, liquidity, and investor participation in the stock.

Analyzing this dataset can help investors, traders, and financial analysts make informed decisions about buying, selling, or holding the company's stock based on historical price movements and trading metrics.